# shop/api_specs/api_spec.json
{
    "swagger": "2.0",
    "host": "127.0.0.1:8000",
    "basePath": "/api/shop/",
    "info": {
        "version": "1.0.0",
        "title": "Simple API",
        "description": "A simple API to learn how to write OpenAPI Specification"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "security": [],
    "securityDefinitions": {
      "oauth": {
          "tokenUrl": "http://auth.ibtspl.com/oauth2/",
          "flow": "password",
          "scopes": {
              "read": "read users",
              "write": "create users",
              "update": "update users",
              "delete": "delete users",
              "superuser": "super user permission" 
          },
          "type": "oauth2"
      }
  },
    "definitions": {
...
        "ProductId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "BasicProduct": {
            "type": "object",
            "properties": {
                "name":{
                  "type":"string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format":"float"
                },
                "mfg_date": {
                    "type": "string"
                },
                "exp_date": {
                    "type": "string"
                },
                "category":{
                     "type":"string"
                },
                "stock_quantity": {
                    "type": "number",
                    "format":"int32"
                }
            },
            "required": [
                "description",
                "name",
                "stock_quantity",
                "exp_date",
                "mfg_date",
                "price"
            
            ]
        },
        "Product": {
            "allOf": [
                {
                    "$ref": "#/definitions/BasicProduct"
                },
                {
                    "$ref": "#/definitions/ProductId"
                }
            ]
        },
...
        "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" }
        }
      },
...
      "Payment": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "order_id": { "type": "string" },
          "amount": { "type": "number", "format": "float"},
          "method": { "type": "string", "enum": ["Card", "NetBanking", "UPI", "COD"] },
          "status": { "type": "string", "enum": ["Success", "Pending", "Failed"] },
          "transaction_date": { "type": "string", "format": "date-time" }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "products": {
            "type": "array",
            "items": { "$ref": "#/definitions/Product" }
          },
          "customer": { "$ref": "#/definitions/Customer" },
          "payment": { "$ref": "#/definitions/Payment" },
          "status": { "type": "string", "enum": ["Pending", "Shipped", "Delivered", "Cancelled"] },
          "expected_delivery_date": { "type": "string", "format": "date" },
          "order_datetime": { "type": "string", "format": "date-time" }
        }
      },
      "OrderList": {
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": { "$ref": "#/definitions/Order" }
          },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "size": { "type": "integer" }
        }
      },
...
    },
...
    "paths": {
...
"/products/": {
    "get": {
      "summary": "List all products",
      "description": "Retrieve a paginated list of all available products.",
      "operationId": "get_products",
      "security": [
                    {
                        "oauth": ["read"]
                    }
                ],
      "parameters": [
        {
          "name": "limit",
          "in": "query",
          "description": "The number of results per page for paginated results.",
          "required": true,
          "type": "integer",
          "minimum": 1
        },
        {
          "name": "offset",
          "in": "query",
          "description": "The starting point of the page of results to retrieve (e.g., offset = page_number * limit).",
          "required": true,
          "type": "integer",
          "minimum": 0
        }
      ],
      "responses": {
        "200": {
          "description": "A paginated list of products.",
          "schema": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    }
            }, 
        "/products/search/": {
    "get": {
      "summary": "Search products",
      "description": "Search for products using name or category as query parameters.",
      "operationId": "search_products",
      "security": [
                    {
                        "oauth": ["read"]
                    }
                ],
      "parameters": [
        {
          "name": "name",
          "in": "query",
          "description": "The name of the product to search for.",
          "required": false,
          "type": "string"
        },
        {
          "name": "category",
          "in": "query",
          "description": "The category of the product to search for.",
          "required": false,
          "type": "string"
        }
      ],
      "responses": {
        "200": {
          "description": "A list of products matching the search criteria.",
          "schema": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    }
  }, 
...
          "/products/low/stock/": {
        "get": {
          "summary": "Get low-stock alerts",
          "description": "Retrieve a list of products with stock below a defined threshold.",
          "operationId":"get_product_low_stock",
          "security": [
                    {
                        "oauth": ["read"]
                    }
                ],
          "responses": {
            "200": {
              "description": "A list of low-stock products.",
              "schema": {
              "type": "array",
              "items": {
                  "$ref": "#/definitions/Product"
                      }
           }
            }
          }
        }
    },
    "/customers/": {
        "get": {
          "summary": "List all customers",
          "description": "Retrieve a paginated list of all available customers.",
          "operationId": "get_customers",
          "security": [
                    {
                        "oauth": ["read"]
                    }
                ],
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "description": "The number of results per page for paginated results.",
              "required": false,
              "type": "integer",
              "minimum": 1
            },
            {
              "name": "offset",
              "in": "query",
              "description": "The starting point of the page of results to retrieve (e.g., offset = page_number * limit).",
              "required": false,
              "type": "integer",
              "minimum": 0
            }
          ],
          "responses": {
            "200": {
              "description": "A paginated list of customers.",
              "schema": {
                "$ref": "#/definitions/Customer"
              }
            }
          }
        }
                },
...
               "/customers/update/{id}/": {
            "put": {
              "operationId": "update_customer",
              "summary": "Update customer details",
              "description": "Modify customer information such as address or phone number.",
              "security": [
                    {
                        "oauth": ["update"]
                    }
                ],
              "parameters": [
                { "name": "id", "in": "path", "required": true, "type":"string"  }
              ],
              "responses": {
                "200": { "description": "Customer successfully updated." }
              }
            }
               },
...
      "/orders/": {
  "get": {
    "summary": "List all orders",
    "description": "Retrieve a paginated list of all orders placed by customers.",
    "operationId": "list_orders",
    "security": [
                    {
                        "oauth": ["read"]
                    }
                ],
    "parameters": [
      {
        "name": "page",
        "in": "query",
        "description": "The page number to retrieve.",
        "required": false,
        "type": "integer",
        "minimum": 1
      },
      {
        "name": "size",
        "in": "query",
        "description": "The number of orders per page.",
        "required": false,
        "type": "integer",
        "minimum": 1
      }
    ],
    "responses": {
      "200": {
        "description": "A paginated list of orders.",
        "schema": {
          "$ref": "#/definitions/OrderList"
        }
      }
    }
  }
},
...
"/payments/": {
  "get": {
    "summary": "List all payments",
    "description": "Retrieve a paginated list of all payments made by customers.",
    "operationId": "get_payments",
    "security": [
                    {
                        "oauth": ["read"]
                    }
                ],
    "parameters": [
      {
        "name": "page",
        "in": "query",
        "description": "The page number to retrieve for paginated results.",
        "required": false,
        "type": "integer",
        "minimum": 1,
        "default": 1
      },
      {
        "name": "size",
        "in": "query",
        "description": "The number of results per page for paginated results.",
        "required": false,
        "type": "integer",
        "minimum": 1,
        "default": 10
      }
    ],
    "responses": {
      "200": {
        "description": "A paginated list of payments.",
        "schema": {
          "type": "object",
          "properties": {
            "page": {
              "type": "integer",
              "example": 1
            },
            "size": {
              "type": "integer",
              "example": 10
            },
            "totalPages": {
              "type": "integer",
              "example": 5
            },
            "totalPayments": {
              "type": "integer",
              "example": 50
            },
            "payments": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Payment"
              }
            }
          }
        }
      },
      "400": {
        "description": "Invalid pagination parameters.",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "example": "Page number must be greater than zero."
            }
          }
        }
      }
    }
  }
},
...
}
}